break prevote equivocation test into its own file
